+++ R/esp-get-railway.R
@@ -18,7 +18,6 @@
 #'
 #' @examplesIf esp_check_access()
 #' \donttest{
-#'
 #' provs <- esp_get_prov()
 #' ccaa <- esp_get_ccaa()
 #'
@@ -36,20 +35,16 @@
 #'   geom_sf(fill = "grey99", color = "grey50") +
 #'   geom_sf(data = ccaa, fill = NA) +
 #'   geom_sf(
-#'     data = rails, aes(color = tipo),
-#'     show.legend = FALSE, linewidth = 1.5
+#'     data = rails, aes(color = t_ffcc_desc),
+#'     show.legend = FALSE,
+#'     linewidth = 1.5
 #'   ) +
 #'   geom_sf(
 #'     data = stations,
 #'     color = "red", alpha = 0.5
 #'   ) +
-#'   coord_sf(
-#'     xlim = c(-7.5, -2.5),
-#'     ylim = c(38, 41)
-#'   ) +
-#'   scale_color_manual(values = hcl.colors(
-#'     length(unique(rails$tipo)), "viridis"
-#'   )) +
+#'   scale_colour_viridis_d() +
+#'   facet_wrap(~t_ffcc_desc) +
 #'   theme_minimal()
 #' }
 esp_get_railway <- function(
@@ -115,8 +110,29 @@ esp_get_railway <- function(
     }
   }
 
-  data_sf <- sf::st_transform(data_sf, as.double(init_epsg))
+  # Add descriptions
+  # Acceso
+  acc <- db_valores[db_valores$campo == "tipoffcc", 2:3]
+  names(acc) <- c("t_ffcc", "t_ffcc_desc")
+  data_sf <- merge(data_sf, acc, all.x = TRUE)
 
+  # Estado fisico
+  est <- db_valores[db_valores$campo == "estadofisico", 2:3]
+  names(est) <- c("estado_fis", "estado_fis_desc")
+  data_sf <- merge(data_sf, est, all.x = TRUE)
+
+  # Ancho Vias
+  acc <- db_valores[db_valores$campo == "anchovia", 2:3]
+  names(acc) <- c("ancho_via", "ancho_via_desc")
+  data_sf <- merge(data_sf, acc, all.x = TRUE)
+
+  # Numero de vias
+  tip <- db_valores[db_valores$campo == "numerovias", 2:3]
+  names(tip) <- c("num_vias", "num_vias_desc")
+  data_sf <- merge(data_sf, tip, all.x = TRUE)
+
+  data_sf <- sanitize_sf(data_sf)
+  data_sf <- data_sf[order(data_sf$t_ffcc, data_sf$ancho_via), ]
   data_sf <- sf::st_transform(data_sf, as.double(init_epsg))
   data_sf
 }