+++ data-raw/sysdata.R
@@ -27,7 +27,29 @@ esp_grid_ccaa <- mapSpain:::read_geo_file_sf(
   mapSpain:::sanitize_sf()
 
 
+# Databases
+dbs <- list.files(
+  "data-raw/listas_de_valores_enumerados/",
+  pattern = "dbf",
+  full.names = TRUE
+)
+
+
+db_valores <- lapply(dbs, function(x) {
+  f <- tibble::as_tibble(foreign::read.dbf(x))
+  clean_x <- gsub("lve_", "", basename(x))
+  clean_x <- gsub(".dbf", "", clean_x)
+  clean_x <- tolower(clean_x)
+  f$campo <- clean_x
+  f[, unique(c("campo", colnames(f)))]
+}) |>
+  dplyr::bind_rows() |>
+  dplyr::as_tibble() |>
+  dplyr::mutate(descrip = as.character(descrip))
+
+
 usethis::use_data(
+  db_valores,
   esp_hexbin_ccaa,
   esp_hexbin_prov,
   esp_grid_ccaa,